requirements.txt
paho-mqtt >= 1.6.0
bleak >= 0.20.0
threading